Clean up some stuff and get loop animation component almost working again with avatar animations
